+++ packages/terra-form-input/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* maxWidth prop added to pass through to terra-form-field
 
 1.31.1 - (November 20, 2018)
 ------------------